+++ lib/bibliothecary/parsers/bazel.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Bibliothecary
   module Parsers
     class Bazel
@@ -23,7 +25,6 @@ module Bibliothecary
       DEPENDENCY_VERSION = /(?:^|[,(]\s*)version\s*=\s*(?<quote>["'])(?<value>(?:\\.|(?!\k<quote>).)*)\k<quote>/m
       DEPENDENCY_TYPE     = /(?:^|[,(]\s*)dev_dependency\s*=\s*(?<value>True|False)\b/m
 
-
       def self.file_patterns
         ["MODULE.bazel"]
       end
@@ -40,24 +41,21 @@ module Bibliothecary
       def self.parse_module_bazel(file_contents, options: {})
         source = options.fetch(:filename, nil)
 
-        dependencies = file_contents.scan(BAZEL_DEP_STATEMENT).map do |(statement)|
-          name = statement.match(DEPENDENCY_NAME)[:value]
+        dependencies = file_contents.scan(BAZEL_DEP_STATEMENT).filter_map do |(statement)|
+          name_match = statement.match(DEPENDENCY_NAME)
+          next unless name_match
+
           parsed_version = statement.match(DEPENDENCY_VERSION)
           parsed_type = statement.match(DEPENDENCY_TYPE)
-          version = parsed_version ? parsed_version[:value] : '*'
-          type =
-            if parsed_type && parsed_type[:value] == "True"
-              "development"
-            else
-              "runtime"
-            end
+          version = parsed_version ? parsed_version[:value] : "*"
+          type = parsed_type && parsed_type[:value] == "True" ? "development" : "runtime"
 
           Dependency.new(
             platform: platform_name,
-            name: name,
+            name: name_match[:value],
             requirement: version,
             type: type,
-            source: source,
+            source: source
           )
         end
         ParserResult.new(dependencies: dependencies)